# HAL_ChibiOS: optimisation for AP_HAL::micros() with 32 bit timers