--- a/libraries/AP_HAL/system.cpp
+++ b/libraries/AP_HAL/system.cpp
@@ -5,6 +5,11 @@ uint16_t WEAK AP_HAL::millis16()
     return millis() & 0xFFFF;
 }
 
+uint16_t WEAK AP_HAL::micros16()
+{
+    return micros() & 0xFFFF;
+}
+
 void WEAK AP_HAL::dump_stack_trace()
 {
     // stack dump not available on this platform