--- a/libraries/AP_HAL_ChibiOS/RCOutput.h
+++ b/libraries/AP_HAL_ChibiOS/RCOutput.h
@@ -332,7 +332,7 @@ private:
         // serial output
         struct {
             // expected time per bit
-            uint32_t bit_time_us;
+            uint16_t bit_time_us;
 
             // channel to output to within group (0 to 3)
             uint8_t chan;
@@ -411,7 +411,7 @@ private:
         ioline_t line;
 
         // time the current byte started
-        uint32_t byte_start_tick;
+        uint16_t byte_start_tick;
 
         // number of bits we have read in this byte
         uint8_t nbits;
@@ -423,7 +423,7 @@ private:
         uint16_t byteval;
 
         // expected time per bit in micros
-        uint32_t bit_time_tick;
+        uint16_t bit_time_tick;
 
         // the bit value of the last bit received
         uint8_t last_bit;