most ChibiOS boards have a 32 bit system timer. This optimisation reduces the cost of AP_HAL::micros() from 0.3us to 0.06us, which is significant in interrupt handlers and for accurate timing. It takes advantage of the timer being 32 bit with 1MHz clock
This may help for R/C decoding and for blheli pass-thru parsing
Speed testing:
 - H7 (CubeOrange) micros() time drops from 0.3us to 0.06us
 - F404 (KakuteF4) micros16() costs 0.12us, whereas micros() costs 0.82us (16 bit timer)
 - F765 (CUAVv5Nano) old micros() costs 0.58us, new micros() and micros16() cost 0.095

Update: I've added AP_HAL::micros16() which is fast for both 16 bit and 32 bit timers. I've changed blheli pass-thru to use it for the IRQ, which makes the bit timing more accurate. I've tested it on a Wraith32 4-in-1 ESC and successfully updated BLHeli32 fw on all 4 ESCs with no errors with passthru
![image](https://user-images.githubusercontent.com/831867/135705811-f893018b-472a-4799-bfc6-dde128545b2e.png)